% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{finalise_plot}
\alias{finalise_plot}
\title{Add ALA logo to finished plot}
\usage{
finalise_plot(plt, filename = "plot.png", width = 900, height = 600)
}
\arguments{
\item{plt}{ggplot2 object: plot to customise and save}

\item{filename}{string: where to save file}

\item{width}{numeric: width of final plot in pixels}

\item{height}{numeric: height of final plot in pixels}
}
\description{
Add ALA logo to finished plot
}
